engine: extract dispatch::skill_test (#159)
Move the skill-test resolution cluster (start_skill_test,
finish_skill_test, drive_skill_test, validate_commit_indices,
sum_skill_value, sum_committed_icons, resolve_chaos_token_and_emit,
discard_committed_cards, apply_skill_test_follow_up,
fire_on_skill_test_resolution, perform_skill_test, peril_check) into
a new dispatch/skill_test.rs submodule. Pure code move — no logic
changes. Call sites in mod.rs updated to skill_test::<fn>; calls
inside skill_test.rs to helpers still in mod.rs use super::<fn>.
Two intra-doc links pointing at super::close_reaction_window_at and
super::open_queued_reaction_window annotated with explicit link
targets so cargo doc stays clean.
